moved image operations to separate file
